+++ src/main/java/com/authlete/common/package-info.java
@@ -17,11 +17,11 @@
  * <span class="tag">&lt;dependency&gt;
  *     &lt;groupId&gt;</span>com.authlete<span class="tag">&lt;/groupId&gt;
  *     &lt;artifactId&gt;</span>authlete-java-common<span class="tag">&lt;/artifactId&gt;
- *     &lt;version&gt;</span>2.46<span class="tag">&lt;/version&gt;
+ *     &lt;version&gt;</span>2.47<span class="tag">&lt;/version&gt;
  * &lt;/dependency&gt;</span><!--
  * --></pre>
  *
- * @version 2.46
+ * @version 2.47
  */
 package com.authlete.common;